backend/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
databases==0.9.0
exceptiongroup==1.2.1
h11==0.14.0
idna==3.7
Mako==1.3.5
MarkupSafe==2.1.5
Naked==0.1.32
psycopg2-binary==2.9.9
pycryptodome==3.20.0
pydantic==2.7.2
pydantic_core==2.18.3
PyJWT==2.8.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
shellescape==3.8.1
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
starlette-auth-toolkit==0.5.0
starlette_jwt==0.1.9
typing_extensions==4.12.0
urllib3==2.2.1
uvicorn==0.30.0
